Added create user and book method

+++ src/main/java/com/library/system/model/Book.java
@@ -6,14 +6,14 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Book {
-    private Long id;
+    private Integer id;
     private String title;
     private String author;
     private String description;
     private Double price;
 
 
-    public Book(Long id, String title, String author, String description, Double price) {
+    public Book(Integer id, String title, String author, String description, Double price) {
         this.id = id;
         this.title = title;
         this.author = author;
